Move toggleCaption internals to captions object
